Alter ad positions from visibility of site header.
This utilizes checking if the site header is still visible (this could be changed to any element prior to these in the DOM or if the script is moved to the footer or somewhere "lower" in the DOM an element that appears across all pages, I thought the header was a good choice since it should always be there). Once that element is no longer visible the position attribute of the wallpapers is adjusted to make them "sticky" until the element is visible again in which it will then return to it's original position.