Move setting of initiallyExpanded to middleware
Move this logic to the middleware so it can set inittially expanded and the cookie only when that middleware is hit vs expanding it all the time when a cookie isnt present.

The middleware will be moved to specific routes(content/section/home) for now and only set/trigger the expanded version on those routes.