Move auto signup into new onAuthenticationSuccessFormatter
Move where and how the auto daily newsletter signup is happening.  Move it to the newly added onAuthenticationSuccessFormatter functionality that has been added.  This allows you to modify the payload prior to sending it to the omeda rapid ident.  This allows us to add/modifiy the promocode being sent along with adding deyploymentProductId signups if the user does not already have a subscription.