Add onLoginLinkSentFormatter check for website product
Add a check to the onLoginLinkSentFormatter to remove the `registration_meter` promoCode if the user is already signed up for the current rapidIdentificationProductId product.  Meaning they have the current website product in the list of user subscriptions.  This is assuming they have already went through this process. It should only apply to new users to this website product product.