Add newsletterPropmpted position logic to middleware
Add newsletterMiddleware support to set enlPrmpted position(pushdown, inbody)

the setting of the enlPrompted is what ultimatly determins if it should show or not, while the postion(cookie/locals) tells it where.

It will either set initiallyInjected || initiallyExpanded to ture when the conditions are met.  At the moment if initiallyInjected is true initiallyExpanded can not be ture.